+++ modules/core/src/main/java/org/openlmis/core/repository/ProgramRepository.java
@@ -28,7 +28,7 @@ public class ProgramRepository {
         return programMapper.getAllActive();
     }
 
-    public List<Program> getByFacility(Long facilityId) {
+    public List<Program> getByFacility(Integer facilityId) {
         return programMapper.getActiveByFacility(facilityId);
     }
 
@@ -36,13 +36,13 @@ public class ProgramRepository {
         return programMapper.getAll();
     }
 
-    public List<Program> filterActiveProgramsAndFacility(List<RoleAssignment> roleAssignments, Long facilityId) {
+    public List<Program> filterActiveProgramsAndFacility(List<RoleAssignment> roleAssignments, Integer facilityId) {
         String programIds = getCommaSeparatedProgramsForDB(roleAssignments);
         return programSupportedMapper.filterActiveProgramsAndFacility(programIds, facilityId);
     }
 
     private String getCommaSeparatedProgramsForDB(List<RoleAssignment> roleAssignments) {
-        List<Long> programIds = new ArrayList<>();
+        List<Integer> programIds = new ArrayList<>();
         for (RoleAssignment roleAssignment : roleAssignments) {
             programIds.add(roleAssignment.getProgramId());
         }
