+++ modules/core/src/test/java/org/openlmis/core/builder/FacilityBuilder.java
@@ -18,20 +18,20 @@ public class FacilityBuilder {
   public static final Property<Facility, String> code = newProperty();
   public static final Property<Facility, String> name = newProperty();
   public static final Property<Facility, String> type = newProperty();
-  public static final Property<Facility, Long> geographicZoneId = newProperty();
+  public static final Property<Facility, Integer> geographicZoneId = newProperty();
   public static final Property<Facility, Boolean> sdp = newProperty();
   public static final Property<Facility, Boolean> active = newProperty();
   public static final Property<Facility, Date> goLiveDate = newProperty();
   public static final Property<Facility, String> operatedByCode = newProperty();
   public static final Property<Facility, GeographicZone> geographicZone = newProperty();
-  public static Property<Facility, Long> typeId = newProperty();
-  private static Property<Facility, Long> operatedById = newProperty();
+  public static Property<Facility, Integer> typeId = newProperty();
+  private static Property<Facility, Integer> operatedById = newProperty();
   public static final Property<Facility,Boolean> dataReportable= newProperty();
 
   public static final String FACILITY_CODE = "F10010";
   public static final String FACILITY_TYPE_CODE = "warehouse";
-  public static final long FACILITY_TYPE_ID = 1L;
-  public static final long GEOGRAPHIC_ZONE_ID = 2L;
+  public static final Integer FACILITY_TYPE_ID = 1;
+  public static final Integer GEOGRAPHIC_ZONE_ID = 2;
   public static final Instantiator<Facility> defaultFacility = new Instantiator<Facility>() {
 
     @Override
@@ -51,7 +51,7 @@ public class FacilityBuilder {
       facility.setDataReportable(lookup.valueOf(dataReportable, true));
       FacilityOperator operatedBy = new FacilityOperator();
       operatedBy.setCode(lookup.valueOf(operatedByCode, "MoH"));
-      operatedBy.setId(lookup.valueOf(operatedById, 1L));
+      operatedBy.setId(lookup.valueOf(operatedById, 1));
       facility.setOperatedBy(operatedBy);
       facility.setGoLiveDate(lookup.valueOf(goLiveDate, new LocalDate(2013, 10, 10).toDate()));
       facility.setModifiedBy("user");
