+++ modules/core/src/test/java/org/openlmis/core/service/FacilityServiceTest.java
@@ -44,7 +44,7 @@ public class FacilityServiceTest {
 
     @Test
     public void shouldGetFacilityById() throws Exception {
-        Long id = 1L;
+        Integer id = 1;
         when(facilityRepository.getFacility(id)).thenReturn(new Facility());
         Facility facility = facilityService.getFacility(id);
         assertThat(facility, is(new Facility()));
